moved and simplified
